support pretty url
