+++ app/server/routes.js
@@ -1,91 +1,7 @@
-import fs from 'fs'
 import morgan from 'morgan'
 
-import force from './middlewares/args/force'
-import preset from './middlewares/args/preset'
-import project from './middlewares/args/project'
-import size from './middlewares/args/size'
-import src from './middlewares/args/src'
-
-import s3 from 'infrastructure/s3'
-import Media from 'entities/Media'
-
-const mediaHandler = (req, res, next) => {
-  const { force, preset, project, src, mode, height, width } = req._args
-
-  const media = Media.create({
-    preset,
-    project,
-    src,
-    height,
-    width,
-    mode
-  })
-
-  const getMeta = force ?
-    Promise.resolve(null) :
-    s3.meta(media.props.remoteTarget).catch(() => null)
-
-  getMeta.then(meta => {
-    if (meta) {
-      res.set('Content-Type', meta.ContentType);
-      res.set('Content-Length', meta.ContentLength);
-      res.set('Last-Modified', meta.LastModified);
-      res.set('ETag', meta.ETag);
-      res.set('Cache-Control', 'public, max-age=2592000');
-      res.set('Expires', new Date(Date.now() + 2592000000).toUTCString());
-
-      s3.receive(media.props.remoteTarget).pipe(res)
-
-      return
-    }
-
-    const rpc = req.app.get('rpc')
-
-    rpc.request('process-media', {
-      media: media.toJSON(),
-      options: preset.values
-    })
-    .waitFor(media.props.uid)
-    .onResponse(response => {
-      console.log('clearing tmp files...')
-      fs.unlink(media.props.localOriginal, () => {})
-      fs.unlink(media.props.localTarget, () => {})
-      console.log('clear tmp files done')
-
-      if (response.data.succeed) {
-        req._args.force = false
-
-        return mediaHandler(req, res, next)
-      }
-
-      res.set('Cache-Control', 'no-store, no-cache, must-revalidate, proxy-revalidate')
-      res.set('Pragma', 'no-cache')
-      res.set('Expires', '0')
-      res.set('Surrogate-Control', 'no-store')
-      res.status(400).json(response.data)
-    })
-    .send()
-  })
-}
-
-const flow = [
-  project,
-  preset,
-  src,
-  size,
-  force,
-  (req, res, next) => {
-    const { preset, project, src } = req._args
-
-    if (!preset || !project || !src) {
-      return res.sendStatus(400)
-    }
-
-    next()
-  },
-  mediaHandler
-]
+import pretty from './routes/pretty'
+import universal from './routes/universal'
 
 export default app => {
   // devlopment log
@@ -99,8 +15,19 @@ export default app => {
   })
 
   // supported endpoints
-  app.get('/p/:slug/:hash/media', flow)
-  app.get('/p/:slug/media', flow)
+  // universal
+  app.use(universal)
+
+  // pretty
+  app.use(pretty)
+
+  // // pretty
+  // app.get([
+  //   '/p/:slug/:hash/*',
+  //   '/p/:slug/*'
+  // ], (req, res, next) => {
+  //   res.json(req.params)
+  // })
 
   // otherwise
   app.use((error, req, res, next) => {
