+++ package.json
@@ -23,6 +23,7 @@
     "recursive-config": "1.0.1",
     "sharp": "0.19.0",
     "short-hash": "1.0.0",
+    "url-join": "4.0.0",
     "uuid": "3.2.1"
   },
   "devDependencies": {
